requirements.txt
requests==2.31.0
fastapi==0.104.1
firebase_admin
python-dotenv==1.0.0
sentry-sdk
uvicorn
pydantic==1.10.9
requests_toolbelt
